Backend/requirements.txt
web3
fastapi
uvicorn
web3.auto
pydantic
python-dotenv
